+++ app/Domains/GroupType/Controllers/GroupTypeController.php
@@ -24,14 +24,14 @@ class GroupTypeController extends Controller
     public function list(FilterGroupTypeRequest $request)
     {
 
-//        abort_if(!auth()->user()->hasPermissionTo(EnumPermissionGroupType::view_groupTypes->value, 'api'), Response::HTTP_FORBIDDEN, '403 Forbidden');
+        abort_if(!auth()->user()->hasPermissionTo(EnumPermissionGroupType::view_groupTypes->value, 'api'), Response::HTTP_FORBIDDEN, '403 Forbidden');
 
         return  GroupTypeResource::collection($this->groupTypeService->list());
     }
 
     public function delete($id)
     {
-//        abort_if(!auth()->user()->hasPermissionTo(EnumPermissionGroupType::delete_groupType->value, 'api'), Response::HTTP_FORBIDDEN, '403 Forbidden');
+        abort_if(!auth()->user()->hasPermissionTo(EnumPermissionGroupType::delete_groupType->value, 'api'), Response::HTTP_FORBIDDEN, '403 Forbidden');
 
         if($this->groupTypeService->delete($id))
         {
@@ -49,57 +49,15 @@ class GroupTypeController extends Controller
 
     public function findById($id)
     {
-//        abort_if(!auth()->user()->hasPermissionTo(EnumPermissionGroupType::view_groupTypes->value, 'api'), Response::HTTP_FORBIDDEN, '403 Forbidden');
+        abort_if(!auth()->user()->hasPermissionTo(EnumPermissionGroupType::view_groupTypes->value, 'api'), Response::HTTP_FORBIDDEN, '403 Forbidden');
 
         return new GroupTypeResource($this->groupTypeService->findById($id));
     }
 
     public function create(StoreGroupTypeRequest $request)
     {
-        /**
-         * Create Group Type
-         * @OA\Post (
-         *     path="/api/groupType/create",
-         *     tags={"Group Type"},
-         *     @OA\RequestBody(
-         *         @OA\MediaType(
-         *             mediaType="application/json",
-         *             @OA\Schema(
-         *                 @OA\Property(
-         *                      type="object",
-         *                      @OA\Property(
-         *                          property="name",
-         *                          type="string"
-         *                      ),
-         *
-         *                 ),
-         *                 example={
-         *                     "title":"example name",
-         *
-         *                }
-         *             )
-         *         )
-         *      ),
-         *      @OA\Response(
-         *          response=200,
-         *          description="success",
-         *          @OA\JsonContent(
-         *              @OA\Property(property="id", type="number", example=1),
-         *              @OA\Property(property="name", type="string", example="name"),
-         *              @OA\Property(property="updated_at", type="string", example="2021-12-11T09:25:53.000000Z"),
-         *              @OA\Property(property="created_at", type="string", example="2021-12-11T09:25:53.000000Z"),
-         *          )
-         *      ),
-         *      @OA\Response(
-         *          response=400,
-         *          description="invalid",
-         *          @OA\JsonContent(
-         *              @OA\Property(property="msg", type="string", example="fail"),
-         *          )
-         *      )
-         * )
-         */
-//        abort_if(!auth()->user()->hasPermissionTo(EnumPermissionGroupType::create_groupType->value, 'api'), Response::HTTP_FORBIDDEN, '403 Forbidden');
+
+        abort_if(!auth()->user()->hasPermissionTo(EnumPermissionGroupType::create_groupType->value, 'api'), Response::HTTP_FORBIDDEN, '403 Forbidden');
 
         $this->groupTypeService->create($request);
         return response()->json([
@@ -111,7 +69,7 @@ class GroupTypeController extends Controller
     public function update($id, UpdateGroupTypeRequest $request)
     {
 
-//        abort_if(!auth()->user()->hasPermissionTo(EnumPermissionGroupType::edit_groupType->value, 'api'), Response::HTTP_FORBIDDEN, '403 Forbidden');
+        abort_if(!auth()->user()->hasPermissionTo(EnumPermissionGroupType::edit_groupType->value, 'api'), Response::HTTP_FORBIDDEN, '403 Forbidden');
 
        if( $this->groupTypeService->update($id, $request))
        {
